% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SSGLint}
\alias{SSGLint}
\title{Fit nonlinear, 2-way interaction models with spike and slab group lasso}
\usage{
SSGLint(Y, x, xnew = NULL, DFmain = 2, DFint = 2, lambda1 = 1,
  lambda0seq = seq(1, 100, by = 1), a = 1, b = dim(x)[2],
  nFolds = 10, updateSigma = TRUE, M = 10, error = 0.001,
  forceGroups = c())
}
\arguments{
\item{Y}{The outcome to be analyzed}

\item{x}{An n by p matrix of covariates}

\item{xnew}{A new n2 by p matrix of covariates at which the outcome will be predicted}

\item{DFmain}{Degrees of freedom of the splines for each main effect}

\item{DFint}{Degrees of freedom of splines for each interaction pair. The total number of
terms in the interaction will be DFint*DFint}

\item{lambda1}{Prior parameter for the slab component of the prior}

\item{lambda0seq}{Sequence of lambda0 values to consider}

\item{a}{First hyperparameter for the beta prior denoting the prior probability of being in the slab}

\item{b}{Second hyperparameter for the beta prior denoting the prior probability of being in the slab}

\item{nFolds}{The number of folds to run cross validation on}

\item{M}{Positive number less than p indicating how often to update theta and sigma. There is no
need to change this unless trying to optimize computation time}

\item{error}{The l2 norm difference that constitutes when the algorithm has converged}

\item{forceGroups}{A vector containing the indices of any groups you wish to automatically
include in the model and not penalize}

\item{UpdateSigma}{True or False indicating whether to estimate the residual variance}
}
\value{
A list of values containing the predictions at the observed locations, predictions
        at the new locations xnew, a vector indicating which covariates have nonzero main effects,
        a matrix indicating which 2-way interactions are nonzero, and an estimate of the residual
        variation in the model.
}
\description{
This function takes in an outcome and covariates, and fits a semiparametric
regression model such that each covariate has a nonlinear main effect and
each pair of interaction terms has a nonlinear interaction term as well
}
\examples{

## Here we generate 200 samples from 100 covariates
n = 200
n2 = 100
G = 10
x = mvtnorm::rmvnorm(n, sigma=diag(G))
xnew = mvtnorm::rmvnorm(n2, sigma=diag(G))
}
