% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{SSGLnpr}
\alias{SSGLnpr}
\title{Do sparse additive regression with splines using spike and slab group lasso}
\usage{
SSGLnpr(Y, x, xnew = NULL, DF = 2, lambda1 = 0.1,
  lambda0seq = seq(1, 25, by = 1), a = 1, b = dim(x)[2],
  nFolds = 5, updateSigma = TRUE, M = 10, error = 0.001,
  forceGroups = c())
}
\arguments{
\item{Y}{The outcome to be analyzed}

\item{x}{An n by p matrix of covariates}

\item{xnew}{A new n2 by p matrix of covariates at which the outcome will be predicted}

\item{DF}{Degrees of freedom of the splines for each covariate}

\item{lambda1}{Prior parameter for the slab component of the prior}

\item{lambda0seq}{Sequence of lambda0 values to consider}

\item{a}{First hyperparameter for the beta prior denoting the prior probability of being in the slab}

\item{b}{Second hyperparameter for the beta prior denoting the prior probability of being in the slab}

\item{nFolds}{The number of folds to run cross validation on}

\item{M}{Positive number less than p indicating how often to update theta and sigma. There is no
need to change this unless trying to optimize computation time}

\item{error}{The l2 norm difference that constitutes when the algorithm has converged}

\item{forceGroups}{A vector containing the indices of any groups you wish to automatically
include in the model and not penalize}

\item{UpdateSigma}{True or False indicating whether to estimate the residual variance}
}
\value{
A list of values containing the predictions at the observed locations, predictions
        at the new locations xnew, a vector indicating which covariates have nonzero coefficients, 
        an n by p matrix of predicted f(x) values for each covariate, an n2 by p matrix of 
        predicted f(xnew) values for each covariate, and an estimate of the residual
        variation in the model.
}
\description{
This function takes in an outcome and covariates, and fits a semiparametric
regression model with additive functions so that the effect of each covariate is modeled
with a spline basis representation
}
\examples{

## Here we generate 200 samples from 100 covariates
n = 200
n2 = 100
G = 100
x = mvtnorm::rmvnorm(n, sigma=diag(G))
xnew = mvtnorm::rmvnorm(n2, sigma=diag(G))
Ynew = 200 + xnew[,1] + xnew[,2] + 0.6*xnew[,2]^2 + rnorm(n2, sd=1)

modSSGLnpr = SSGLnpr(Y=Y, x=x, xnew = xnew, lambda1=.1, DF=2)
mean((modSSGLnpr$predY - Y)^2)
mean((modSSGLnpr$predYnew - Ynew)^2)
modSSGLnpr$nonzero

plot(xnew[,2], modSSGLnpr$fxnew[,2])
points(xnew[,2], xnew[,2] + 0.6*xnew[,2]^2 - 
         mean(xnew[,2] + 0.6*xnew[,2]^2), col=2)
legend("top", c("Estimated", "Truth"),col=1:2, pch=1)
}
